book-service/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==1.0.0
